refactor(model_profiles): split accumulation from the routing read model
model_profiles.py held two change-reasons in one file: folding a finished
run's outcome into stored history, and deriving the signals routing consults
from that history. Adding a routing signal and changing how an outcome is
accumulated shared a module, a review, and a file claim in the scheduler.

Following the #2350 split of audit_substrate.py, this becomes two owners over
one shared vocabulary:

  model_profiles_storage.py     carriers, I/O, apply_run and its _fold_*/
                                _update_* helpers, backfill, reset, the
                                canonical-ID migration and _merge_* catalogue
  model_profiles_read_model.py  get_*_signal, per-domain fit, observed-cost
                                tie-break, complexity/score stat readers
  model_profiles_identity.py    constants, identity resolution, stat
                                primitives both halves genuinely share

Neither owner imports the other, so a signal is exercisable against a profiles
dict a test builds directly, with no RunOutcome applied to produce it.
model_profiles.py remains a documented re-export facade; every moved function
is byte-identical (verified by AST comparison against HEAD), so no signal
changes what it returns and no derivation was merged into another.

Closes #2467
