+++ src/theforge/coordinator/plan_flow.py
@@ -649,7 +649,7 @@ def _apply_post_plan_dev_checkpoint(
         from theforge.coordinator.audit_substrate import (  # noqa: PLC0415
             load_observed_cost_cohorts,
         )
-        from theforge.model_profiles import load_profiles  # noqa: PLC0415
+        from theforge.model_profiles_storage import load_profiles  # noqa: PLC0415
 
         _model_profiles = load_profiles(config.project_root / ".forge" / "model_profiles.yaml")
         _observed_costs, _ = load_observed_cost_cohorts(config.project_root)
